Move create group dialog into its own component
