+++ src/app/app.module.ts
@@ -5,13 +5,15 @@ import { RouteReuseStrategy } from '@angular/router';
 import { FormsModule } from '@angular/forms';
 import {MatDialogModule} from '@angular/material';
 
-import { AppComponent, CreateGroupDialog } from './app.component';
+import { AppComponent } from './app.component';
 import { AppRoutingModule } from './app-routing.module';
 
 import { HttpClientModule, HTTP_INTERCEPTORS } from '@angular/common/http';
 import { JwtInterceptor, ErrorInterceptor } from '_helpers/index';
 import { AlertService, AuthenticationService } from '_services/index';
 import { ActionSnackBarComponent } from '_components/action-snack-bar/action-snack-bar';
+import { CreateGroupDialog } from '_components/create-group-dialog/create-group-dialog';
+
 
 import { AngularMaterialModule } from '_components/angular-material.module';
 import { ReactiveFormsModule } from '@angular/forms';
